Moved all files into a halibot package
