+++ halibot/__init__.py
@@ -1,0 +1,3 @@
+from .halibot import Halibot
+from .halagent import HalAgent
+from .halmodule import HalModule
