+++ halibot/halobject.py
